Refactor Client
