+++ src/lib.rs
@@ -64,7 +64,7 @@ pub use url::Url;
 
 #[cfg(feature = "mqtt")]
 pub use self::node_api::mqtt::*;
-pub use self::{builder::ClientBuilder, client::*, error::*, utils::*};
+pub use self::{builder::ClientBuilder, client::*, error::*, node_api::core::routes::NodeInfoWrapper, utils::*};
 
 #[cfg(feature = "mqtt")]
 mod async_runtime {
